back/requirements.txt
fastapi == 0.95.2
uvicorn == 0.22.0
pydantic == 1.10.7
sqlalchemy == 1.4.47
alembic == 1.10.3
httpx == 0.24.0
python-multipart == 0.0.6
python-dotenv == 1.0.0
psycopg2-binary == 2.9.6
bcrypt==4.0.1
passlib==1.7.4
python-jose == 3.3.0
email-validator == 2.0.0